Created Context API and Add UserContext for online status and refactor usage
Introduced UserContextApi to manage and provide online status across the app. Refactored EmotionTracker to use the new context instead of local state and event listeners. Wrapped App with UserProvider in main.jsx to enable context usage throughout the component tree.
